Remove `BuiltinLintDiag` usage in `rustc_ast_passes`
